requirements.txt
# Core ML/DL Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# OCR and Document Processing
pytesseract>=0.3.10
pdf2image>=1.16.3
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.10.0

# NLP
nltk>=3.8.1
spacy>=3.6.0
sentencepiece>=0.1.99

# Computer Vision
timm>=0.9.0

# Model Training & Tracking
wandb>=0.15.0
tensorboard>=2.13.0
mlflow>=2.5.0

# Web Framework & UI
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# AWS Integration
boto3>=1.28.0
botocore>=1.31.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Utilities
pyyaml>=6.0
python-json-logger>=2.0.7
tqdm>=4.65.0
requests>=2.31.0
aiohttp>=3.8.5

# Metrics & Evaluation
scikit-metrics>=0.1.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.3

# Production
gunicorn>=21.2.0
prometheus-client>=0.17.0
